chore: add GPLv2 license
The reason I chose GPLv2 and not the GFDL, which is a more logical
choice, is because that's what the Free Pascal team has decided for
their own documentation. As such, it is easier for me to stick to GPLv2
than it is to convince Michaël et al. to switch to GFDL.

Signed-off-by: Ștefan Alecu <165364995+overanalytcl@users.noreply.github.com>